logo moved to common
