+++ b2b-sample/guardio-insurance-sample-application/components/common/logo/logo.js
@@ -18,7 +18,7 @@
 
 import Image from "next/image";
 import React from "react";
-import logoImage from "../../public/logo.png";
+import logoImage from "../../../public/logo.png";
 
 /**
  *
